requirements.txt
grpcio==1.54.0
grpcio-tools==1.54.0
flask==2.3.2
flask-cors==4.0.0